src/requirements.txt
# QuantumViz Agent - Python Dependencies

# AWS SDK and Services
boto3==1.34.0
botocore==1.34.0
aws-cdk-lib==2.131.0
constructs==10.3.0

# Quantum Computing Frameworks
qiskit==0.46.0
qiskit-aer==0.13.0
qiskit-ibm-runtime==0.16.0
cirq==1.3.0
amazon-braket-sdk==1.78.0

# AI and Machine Learning
openai==1.6.1
anthropic==0.7.8
transformers==4.36.0
torch==2.1.0

# Data Processing and Visualization
numpy==1.24.3
scipy==1.11.4
matplotlib==3.7.2
plotly==5.17.0
pandas==2.0.3

# Web Framework and API
flask==2.3.2
flask-cors==4.0.0
pydantic==2.5.0
requests==2.31.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.7.0
